initial tagging and marshaller for extendability
Also includes a handful of fixes to the (renamed) Coder.

Indexing on an int type using riak protobufs seems to be broken since the internals are of type []byte.  The server throws back a pre-commit error.  I'll post to the email list or ask in irc about this at some point.

Next commit will be the Unmarshaller logic.